api/serializers.py
from rest_framework import serializers
from .models import Letter, Word

class WordSerializer(serializers.ModelSerializer):
    class Meta:
        model = Word
        fields = '__all__'

class LetterSerializer(serializers.ModelSerializer):
    words = WordSerializer(many=True, read_only=True)

    class Meta:
        model = Letter
        fields = ['name', 'words']