requirements.txt
dj-database-url==0.5.0
Django==2.2.6
django-cors-headers==3.1.1
django-filter==2.2.0
django-filters==0.2.1
djangorestframework==3.10.3
gunicorn==19.9.0
psycopg2==2.8.3
psycopg2-binary==2.8.3
python-decouple==3.1
pytz==2019.2
sqlparse==0.3.0
virtualenv==16.7.5
whitenoise==4.1.4
